Make random shuffle actually random
Currently the randomness in the random shuffle is implemented by
returning a random result in a sort function, which is not very random
at all. In practice, it results in songs close to the start of the list
being sorted first.

To fix this, we can simply assign a random integer to each song and
compare them in the sorting function, to make the results consistent.

Yes, I know this probably isn't a great way to implement a random
shuffle in a music player app. We would probably want to write our own
list model to handle smarts like "don't sort two songs from the same
artist next to each other." But that requires effort, whereas what I've
done here is simple and works.

Fixes #369